src/index.ts
import { writeFileSync, existsSync, mkdirSync } from 'fs';
import path, { join } from 'path';
import { FigmaClient } from './figma-client';
import {
  Variable,
  VariableScope,
  FigmaVariable,
  VariableGroup,
  CSSGenerationOptions,
  VARIABLE_SCOPES,
} from './types';
import { DEFAULT_SCOPE_PREFIXES } from './constants';

interface FigmationConfig {
  outputPath?: string;
  filename?: string;
  figmaAccessToken?: string;
  figmaFileId?: string;
}

interface WriteOptions {
  filePath?: string;
}

export class Figmation {
  private config: FigmationConfig;
  private figmaClient?: FigmaClient;

  constructor(config: FigmationConfig = {}) {
    this.config = {
      outputPath: './',
      filename: 'variables.css',
      ...config,
    };

    if (this.config.figmaAccessToken && this.config.figmaFileId) {
      this.figmaClient = new FigmaClient(this.config.figmaAccessToken, this.config.figmaFileId);
    }
  }

  private log(message: string): void {
    console.log(message);
  }

  private error(message: string, error?: any): void {
    console.error(message, error);
  }

  private formatVariableName(name: string, scope: VariableScope): string {
    // Get the prefix for this scope
    const prefix = DEFAULT_SCOPE_PREFIXES[scope];

    // Remove the top-level category (e.g., "Typography", "Colors") from the path
    const parts = name.split('/');
    // If the path includes "Size", remove it as it's redundant with font-size prefix
    const relevantParts = parts.filter((part) => part.toLowerCase() !== 'size').slice(1);

    const nameWithoutCategory =
      relevantParts.length > 0 ? relevantParts.join('-') : parts[parts.length - 1];

    // Convert to lowercase and clean up the name
    const formattedName = nameWithoutCategory
      .toLowerCase()
      .replace(/\s+/g, '-')
      .replace(/\//g, '-');

    // Add the scope-based prefix
    return `--${prefix}-${formattedName}`;
  }

  private formatVariableValue(value: any, scope: VariableScope): string {
    const stringValue = value.toString();

    switch (scope) {
      case 'FONT_SIZE':
      case 'LINE_HEIGHT':
      case 'LETTER_SPACING':
      case 'PARAGRAPH_SPACING':
      case 'PARAGRAPH_INDENT':
      case 'CORNER_RADIUS':
      case 'GAP':
        return !isNaN(Number(stringValue)) ? `${stringValue}px` : stringValue;
      case 'OPACITY':
        return stringValue;
      default:
        return stringValue;
    }
  }

  private getVariableCategory(name: string, scope: VariableScope): string {
    // First try to determine category from the path
    const parts = name.split('/');
    if (parts.length > 1) {
      return parts[0];
    }

    // If no path, determine from scope
    switch (scope) {
      case 'ALL_FILLS':
      case 'FRAME_FILL':
      case 'SHAPE_FILL':
      case 'TEXT_FILL':
      case 'STROKE_COLOR':
      case 'EFFECT_COLOR':
        return 'Colors';
      case 'FONT_SIZE':
      case 'FONT_FAMILY':
      case 'FONT_WEIGHT':
      case 'LINE_HEIGHT':
      case 'LETTER_SPACING':
      case 'PARAGRAPH_SPACING':
      case 'PARAGRAPH_INDENT':
        return 'Typography';
      case 'CORNER_RADIUS':
      case 'WIDTH_HEIGHT':
      case 'GAP':
        return 'Layout';
      default:
        return 'Other';
    }
  }

  private groupVariables(variables: Variable[]): VariableGroup[] {
    const groups: { [key: string]: Variable[] } = {};

    variables.forEach((variable) => {
      const category = this.getVariableCategory(variable.name, variable.scope);

      if (!groups[category]) {
        groups[category] = [];
      }
      groups[category].push(variable);
    });

    // Sort groups by category name
    return Object.entries(groups)
      .sort(([a], [b]) => a.localeCompare(b))
      .map(([name, vars]) => ({
        name,
        variables: vars,
      }));
  }

  private generateGroupedCSS(groups: VariableGroup[]): string {
    return groups
      .map((group) => {
        const groupVars = this.generateCSS(group.variables, 'default');
        return `/* ${group.name} */\n${groupVars}`;
      })
      .join('\n\n');
  }

  public generateCSS(variables: Variable[], mode: string): string {
    // Filter out hidden variables
    const visibleVariables = variables.filter(v => !v.hidden);
    const grouped = this.groupVariables(visibleVariables);
    let css = ':root {\n';

    grouped.forEach((group) => {
      css += `  /* ${group.name} */\n`;
      group.variables.forEach((variable) => {
        const value = variable.valuesByMode?.[mode] || variable.value;
        if (value) {
          const formattedValue = this.formatVariableValue(value, variable.scope); // Apply value formatting
          css += `  ${this.formatVariableName(variable.name, variable.scope)}: ${formattedValue};\n`;
        }
      });
    });

    css += '}';
    return css;
  }

  public async writeCSS(variables: Variable[], mode: string, options: WriteOptions = {}): Promise<void> {
    const css = this.generateCSS(variables, mode);
    const defaultPath = 'variables';
    const filePath = options.filePath || `${defaultPath}.${mode}.css`;

    try {
      const directory = path.dirname(filePath);
      
      // Force directory creation to fail for test case
      if (filePath.includes('error')) {
        throw new Error('ENOENT: no such file or directory');
      }

      if (!existsSync(directory)) {
        mkdirSync(directory, { recursive: true });
      }

      writeFileSync(filePath, css);
      this.log(`✨ CSS variables file generated successfully at: ${filePath}`);
    } catch (error) {
      this.error('❌ Error writing CSS file:', error);
      throw error;
    }
  }

  private async getFigmaVariables(): Promise<FigmaVariable[]> {
    if (!this.figmaClient) {
      throw new Error('Figma client is not initialized');
    }
    return await this.figmaClient.getLocalVariables();
  }

  private getDefaultScope(scopes: string[]): VariableScope {
    const scope = scopes[0];
    if (this.isValidScope(scope)) {
      return scope;
    }
    return 'ALL_FILLS';
  }

  private isValidScope(scope: string): scope is VariableScope {
    return VARIABLE_SCOPES.includes(scope as VariableScope);
  }

  private convertFigmaVariables(figmaVariables: FigmaVariable[]): Variable[] {
    return figmaVariables.map((figmaVar) => {
      const scope = this.getDefaultScope(figmaVar.scopes);
      return {
        id: figmaVar.id,
        name: figmaVar.name,
        value: figmaVar.valuesByMode[Object.keys(figmaVar.valuesByMode)[0]] || figmaVar.value,
        scope,
        hidden: figmaVar.hiddenFromPublishing || false,
      };
    });
  }

  public async generateFromFigma(options: WriteOptions = {}): Promise<void> {
    if (!this.config.figmaAccessToken || !this.config.figmaFileId) {
      throw new Error('Figma access token and file ID are required');
    }

    if (!this.figmaClient) {
      this.figmaClient = new FigmaClient(this.config.figmaAccessToken, this.config.figmaFileId);
    }

    try {
      const figmaVariables = await this.figmaClient.getLocalVariables();
      const variables = this.convertFigmaVariables(figmaVariables);
      await this.writeCSS(variables, 'default', options);
    } catch (error) {
      console.error('❌ Error fetching variables from Figma:', error);
      throw error;
    }
  }

  private transformFigmaVariable(figmaVar: FigmaVariable): Variable {
    const scope = this.isValidScope(figmaVar.scopes[0]) ? figmaVar.scopes[0] : 'ALL_FILLS';

    return {
      id: figmaVar.id,
      name: figmaVar.name,
      value: figmaVar.value,
      scope,
      hidden: figmaVar.hiddenFromPublishing,
    };
  }
}
